requirements.txt
altgraph==0.17.5
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.4.22
colorama==0.4.6
distro==1.9.0
et_xmlfile==2.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.13
jiter==0.14.0
openai==2.32.0
openpyxl==3.1.5
packaging==26.1
pefile==2024.8.26
pillow==12.2.0
pydantic==2.12.5
pydantic_core==2.41.5
pyinstaller==6.19.0
pyinstaller-hooks-contrib==2026.4
PyMuPDF==1.27.2.2
python-dotenv==1.2.2
pywin32-ctypes==0.2.3
PyYAML==6.0.3
setuptools==82.0.1
sniffio==1.3.1
tqdm==4.67.3
typing-inspection==0.4.2
typing_extensions==4.15.0
xlsxwriter==3.2.9
